requirements.txt
asgiref==3.5.0
boto3==1.24.7
botocore==1.27.7
dj-database-url==0.5.0
Django==4.0.3
django-admin-thumbnails==0.2.6
django-ckeditor==6.2.0
django-credit-cards==0.4.1
django-heroku==0.3.1
django-js-asset==2.0.0
django-mptt==0.13.4
django-storages==1.12.3
django-widget-tweaks==1.4.12
jmespath==1.0.0
numpy==1.22.3
opencv-contrib-python==4.5.5.64
opencv-python==4.5.5.64
Pillow==9.0.1
psycopg2==2.9.3
python-dateutil==2.8.2
python-decouple==3.6
s3transfer==0.6.0
six==1.16.0
sqlparse==0.4.2
tzdata==2022.1
urllib3==1.26.9
whitenoise==6.0.0
